==> 2_bdd_old.ipynb <==
import pandas as pd
import glob
import numpy as np
import re

df6=pd.read_excel('C:/Users/d.mandon/Imadis Téléradiologie/INTRANET - IMADIS/QUALITE/7- RHM/15 - DMA/GitHub/data/2020.21_AO.xlsx')
df7=pd.read_excel('C:/Users/d.mandon/Imadis Téléradiologie/INTRANET - IMADIS/QUALITE/7- RHM/15 - DMA/GitHub/data/2020.21_CDS.xlsx')
df8=pd.read_excel('C:/Users/d.mandon/Imadis Téléradiologie/INTRANET - IMADIS/QUALITE/7- RHM/15 - DMA/GitHub/data/2020.21_PDS.xlsx')
df9=pd.read_excel('C:/Users/d.mandon/Imadis Téléradiologie/INTRANET - IMADIS/QUALITE/7- RHM/15 - DMA/GitHub/data/2020.21_RRU.xlsx')


df10=pd.read_excel('C:/Users/d.mandon/Imadis Téléradiologie/INTRANET - IMADIS/QUALITE/7- RHM/15 - DMA/GitHub/data/2021-X_AO.xlsx')
df11=pd.read_excel('C:/Users/d.mandon/Imadis Téléradiologie/INTRANET - IMADIS/QUALITE/7- RHM/15 - DMA/GitHub/data/2021-X_ART.xlsx')
df12=pd.read_excel('C:/Users/d.mandon/Imadis Téléradiologie/INTRANET - IMADIS/QUALITE/7- RHM/15 - DMA/GitHub/data/2021-X_CDS.xlsx')
df13=pd.read_excel('C:/Users/d.mandon/Imadis Téléradiologie/INTRANET - IMADIS/QUALITE/7- RHM/15 - DMA/GitHub/data/2021-X_PDS.xlsx')
df14=pd.read_excel('C:/Users/d.mandon/Imadis Téléradiologie/INTRANET - IMADIS/QUALITE/7- RHM/15 - DMA/GitHub/data/2021-X_RRU.xlsx')


df6.sample(4)

correspondance_colonnes = {'Source.Name': 'Source',
                          'Subject': 'Nom_Prenom',
                          'Start Date': 'Date',
                          'Start Time' : 'Heure_Début',
                          'End Date' : 'Date_Fin',
                          'End Time' : 'Heure_Fin',
                          'All Day Event' : 'Evenement',
                          'Description': 'Horaire',
                          'Location' : 'Type',
                          'Private' : 'Private'}

df6 = df6.rename(columns=correspondance_colonnes)
df7 = df7.rename(columns=correspondance_colonnes)
df8 = df8.rename(columns=correspondance_colonnes)
df9 = df9.rename(columns=correspondance_colonnes)



df6[['Heure_Début', 'Heure_Fin']] = df6['Horaire'].str.split(' - ', expand=True)
df7[['Heure_Début', 'Heure_Fin']] = df7['Horaire'].str.split(' - ', expand=True)
df8[['Heure_Début', 'Heure_Fin']] = df8['Horaire'].str.split(' - ', expand=True)
df9[['Heure_Début', 'Heure_Fin']] = df9['Horaire'].str.split(' - ', expand=True)

df8.head(4)

df6['Date_Heure_Début'] = df6['Date'].dt.strftime('%Y-%m-%d')+' '+df6['Heure_Début']
df7['Date_Heure_Début'] = df7['Date'].dt.strftime('%Y-%m-%d')+' '+df7['Heure_Début']
df8['Date_Heure_Début'] = df8['Date'].dt.strftime('%Y-%m-%d')+' '+df8['Heure_Début']
df9['Date_Heure_Début'] = df9['Date'].dt.strftime('%Y-%m-%d')+' '+df9['Heure_Début']

df6['Date_Heure_Début'] = pd.to_datetime(df6['Date_Heure_Début'], format="%Y/%m/%d %H:%M:%S")
df6['Heure_Début'] = pd.to_datetime(df6['Heure_Début'])
df6['Heure_Début'] = df6['Date_Heure_Début'].dt.strftime('%H:%M:%S')
df6['Heure_Fin'] = pd.to_datetime(df6['Heure_Fin'])
df6['Heure_Fin'] = df6['Heure_Fin'].dt.strftime('%H:%M:%S')

# Créer une nouvelle colonne "Date_Heure_Fin" en utilisant les informations des colonnes existantes
df6['Date_Fin'] = df6.apply(lambda row: row['Date_Heure_Début'].date() if row['Heure_Début'] <= row['Heure_Fin']
                                else (row['Date_Heure_Début'] + pd.DateOffset(days=1)).date(), axis=1)


df7['Date_Heure_Début'] = pd.to_datetime(df7['Date_Heure_Début'], format="%Y/%m/%d %H:%M:%S")
df7['Heure_Début'] = pd.to_datetime(df7['Heure_Début'])
df7['Heure_Début'] = df7['Date_Heure_Début'].dt.strftime('%H:%M:%S')
df7['Heure_Fin'] = pd.to_datetime(df7['Heure_Fin'])
df7['Heure_Fin'] = df7['Heure_Fin'].dt.strftime('%H:%M:%S')

# Créer une nouvelle colonne "Date_Heure_Fin" en utilisant les informations des colonnes existantes
df7['Date_Fin'] = df7.apply(lambda row: row['Date_Heure_Début'].date() if row['Heure_Début'] <= row['Heure_Fin']
                                else (row['Date_Heure_Début'] + pd.DateOffset(days=1)).date(), axis=1)

df8['Date_Heure_Début'] = pd.to_datetime(df8['Date_Heure_Début'], format="%Y/%m/%d %H:%M:%S")
df8['Heure_Début'] = pd.to_datetime(df8['Heure_Début'])
df8['Heure_Début'] = df8['Date_Heure_Début'].dt.strftime('%H:%M:%S')
df8['Heure_Fin'] = pd.to_datetime(df8['Heure_Fin'])
df8['Heure_Fin'] = df8['Heure_Fin'].dt.strftime('%H:%M:%S')

# Créer une nouvelle colonne "Date_Heure_Fin" en utilisant les informations des colonnes existantes
df8['Date_Fin'] = df8.apply(lambda row: row['Date_Heure_Début'].date() if row['Heure_Début'] <= row['Heure_Fin']
                                else (row['Date_Heure_Début'] + pd.DateOffset(days=1)).date(), axis=1)

df9['Date_Heure_Début'] = pd.to_datetime(df9['Date_Heure_Début'], format="%Y/%m/%d %H:%M:%S")
df9['Heure_Début'] = pd.to_datetime(df9['Heure_Début'])
df9['Heure_Début'] = df9['Date_Heure_Début'].dt.strftime('%H:%M:%S')
df9['Heure_Fin'] = pd.to_datetime(df9['Heure_Fin'])
df9['Heure_Fin'] = df9['Heure_Fin'].dt.strftime('%H:%M:%S')

# Créer une nouvelle colonne "Date_Heure_Fin" en utilisant les informations des colonnes existantes
df9['Date_Fin'] = df9.apply(lambda row: row['Date_Heure_Début'].date() if row['Heure_Début'] <= row['Heure_Fin']
                                else (row['Date_Heure_Début'] + pd.DateOffset(days=1)).date(), axis=1)

df6['Date_Fin'] = pd.to_datetime(df6['Date_Fin'])
df7['Date_Fin'] = pd.to_datetime(df7['Date_Fin'])
df8['Date_Fin'] = pd.to_datetime(df8['Date_Fin'])
df9['Date_Fin'] = pd.to_datetime(df9['Date_Fin'])

df6['Date_Heure_Fin'] = df6['Date_Fin'].dt.strftime('%Y-%m-%d')+' '+df6['Heure_Fin']
df7['Date_Heure_Fin'] = df7['Date_Fin'].dt.strftime('%Y-%m-%d')+' '+df7['Heure_Fin']
df8['Date_Heure_Fin'] = df8['Date_Fin'].dt.strftime('%Y-%m-%d')+' '+df8['Heure_Fin']
df9['Date_Heure_Fin'] = df9['Date_Fin'].dt.strftime('%Y-%m-%d')+' '+df9['Heure_Fin']

df6['Associé'] = np.nan
df7['Associé'] = np.nan
df8['Associé'] = np.nan
df9['Associé'] = np.nan

df6['mail'] = np.nan
df7['mail'] = np.nan
df8['mail'] = np.nan
df9['mail'] = np.nan

df6['Site'] = np.nan
df7['Site'] = np.nan
df8['Site'] = np.nan
df9['Site'] = np.nan

df6['Jour/Nuit'] = np.nan
df7['Jour/Nuit'] = np.nan
df8['Jour/Nuit'] = np.nan
df9['Jour/Nuit'] = np.nan

df6 = df6.drop(['Type', 'Private'], axis=1)
df7 = df7.drop(['Type', 'Private'], axis=1)
df8 = df8.drop(['Type', 'Private'], axis=1)
df9 = df9.drop(['Type', 'Private'], axis=1)

df6 = df6.dropna(subset=['Nom_Prenom'])
df7 = df7.dropna(subset=['Nom_Prenom'])
df8 = df8.dropna(subset=['Nom_Prenom'])
df9 = df9.dropna(subset=['Nom_Prenom'])

df6.tail()

df11.tail()

correspondance_colonnes = {'Source.Name': 'Source',
                          'Column1': 'Date',
                          'Column2': 'Associé',
                          'Column3' : 'Nom',
                          'Column4' : 'Prénom',
                          'Column5' : 'mail',
                          'Column6' : 'Site',
                          'Column7': 'Jour/Nuit',
                          'Column8' : 'Date_Heure_Début',
                          'Column9' : 'Date_Heure_Fin',
                          'Column10' : 'Chef',
                          'Column11' : 'Formation'}

df10 = df10.rename(columns=correspondance_colonnes)
df11 = df11.rename(columns=correspondance_colonnes)
df12 = df12.rename(columns=correspondance_colonnes)
df13 = df13.rename(columns=correspondance_colonnes)
df14 = df14.rename(columns=correspondance_colonnes)

df10['Nom_Prenom'] = df10['Nom']+' '+df10['Prénom']
df11['Nom_Prenom'] = df11['Nom']+' '+df11['Prénom']
df12['Nom_Prenom'] = df12['Nom']+' '+df12['Prénom']
df13['Nom_Prenom'] = df13['Nom']+' '+df13['Prénom']
df14['Nom_Prenom'] = df14['Nom']+' '+df14['Prénom']

df10 = df10.drop(['Nom', 'Prénom'], axis=1)
df11 = df11.drop(['Nom', 'Prénom'], axis=1)
df12 = df12.drop(['Nom', 'Prénom'], axis=1)
df13 = df13.drop(['Nom', 'Prénom'], axis=1)
df14 = df14.drop(['Nom', 'Prénom'], axis=1)

df11['Date_Heure_Début'] = pd.to_datetime(df11['Date_Heure_Début'], format="%d/%m/%Y %H:%M:%S")

df10['Heure_Début'] = df10['Date_Heure_Début'].dt.strftime('%H:%M:%S')
df12['Heure_Début'] = df12['Date_Heure_Début'].dt.strftime('%H:%M:%S')
df13['Heure_Début'] = df13['Date_Heure_Début'].dt.strftime('%H:%M:%S')
df14['Heure_Début'] = df14['Date_Heure_Début'].dt.strftime('%H:%M:%S')
df11['Heure_Début'] = df11['Date_Heure_Début'].dt.strftime('%H:%M:%S')

df10['Heure_Fin'] = df10['Date_Heure_Fin'].dt.strftime('%H:%M:%S')
df12['Heure_Fin'] = df12['Date_Heure_Fin'].dt.strftime('%H:%M:%S')
df13['Heure_Fin'] = df13['Date_Heure_Fin'].dt.strftime('%H:%M:%S')
df14['Heure_Fin'] = df14['Date_Heure_Fin'].dt.strftime('%H:%M:%S')
df11['Heure_Fin'] = df11['Date_Heure_Fin'].dt.strftime('%H:%M:%S')

df10['Horaire'] = df10['Heure_Début']+' - '+df10['Heure_Fin']
df11['Horaire'] = df11['Heure_Début']+' - '+df11['Heure_Fin']
df12['Horaire'] = df12['Heure_Début']+' - '+df12['Heure_Fin']
df13['Horaire'] = df13['Heure_Début']+' - '+df13['Heure_Fin']
df14['Horaire'] = df14['Heure_Début']+' - '+df14['Heure_Fin']

df10 = df10.dropna(subset=['Nom_Prenom'])
df11 = df11.dropna(subset=['Nom_Prenom'])
df12 = df12.dropna(subset=['Nom_Prenom'])
df13 = df13.dropna(subset=['Nom_Prenom'])
df14 = df14.dropna(subset=['Nom_Prenom'])

df12.tail()

dfs = [df6, df7, df8, df9, df10, df11, df12, df13, df14]

# Concaténez les DataFrames de la liste en utilisant la fonction concat()
concatenated_df = pd.concat(dfs)

# Réinitialisez l'index du DataFrame concaténé
concatenated_df.reset_index(drop=True, inplace=True)

concatenated_df.tail(10)

#concatenated_df['Date'] = pd.to_datetime(concatenated_df['Date'])
#concatenated_df['Date'] = concatenated_df['Date'].dt.strftime("%d/%m/%Y")

concatenated_df.tail(5)

def extract_value(source):
    result = re.search(r'_(.*?)\.', source)
    if result:
        return result.group(1)
    else:
        return ""

concatenated_df['Source'] = concatenated_df['Source'].apply(extract_value)

concatenated_df=concatenated_df.drop(['Date_Fin','Evenement'], axis=1)

#modifier filtre avec date de fin du dernier tour
df_filtered = concatenated_df.loc[concatenated_df['Date'] <= '2023-07-10']


df_filtered.tail(3)

df_filtered.to_csv('C:/Users/d.mandon/Imadis Téléradiologie/INTRANET - IMADIS/QUALITE/7- RHM/15 - DMA/GitHub/data/Concat_df.csv', index=False)
