==> 6_Création_BDD_SQL.ipynb <==
import sqlite3
import pandas as pd

conn = sqlite3.connect('BDD_PBI.db')


df= pd.read_csv('Concat_df.csv')

Archive_Planning = 'Archive_Planning'
df.to_sql(Archive_Planning, conn, if_exists='replace', index=False)

df= pd.read_csv('new_data.csv')

New_Data_Planning = 'New_Data_Planning'
df.to_sql(New_Data_Planning, conn, if_exists='replace', index=False)

df= pd.read_csv('BDD.csv')

BDD = 'BDD'
df.to_sql(BDD, conn, if_exists='replace', index=False)

conn.close()
